+++ tests/test_board.py
@@ -1,0 +1,104 @@
+import unittest
+from chess.board import Board
+from chess.rook import Rook
+from chess.pawn import Pawn
+from chess.exceptions import OutOfBoard
+
+
+class TestBoard(unittest.TestCase):
+
+    def test_str_board(self):                #Inicializa el tablero y compara la salida esperada con la real utilizando simbolos
+        board = Board()
+        self.assertEqual(
+            str(board),
+            (
+                "♖♘♗♕♔♗♘♖\n"    
+                "♙♙♙♙♙♙♙♙\n"          
+                "        \n"
+                "        \n"            
+                "        \n"
+                "        \n"
+                "♟♟♟♟♟♟♟♟\n"
+                "♜♞♝♛♚♝♞♜\n"
+            )
+        )
+
+    def setUp(self):
+        self.board = Board()
+
+    def test_initial_positions(self):
+        # Verificar que las piezas están en las posiciones correctas
+        self.assertIsInstance(self.board.get_piece(0, 0), Rook)
+        self.assertEqual(self.board.get_piece(0, 0).get_color, "BLACK")
+        self.assertIsInstance(self.board.get_piece(7, 0), Rook)
+        self.assertEqual(self.board.get_piece(7, 0).get_color, "WHITE")
+        self.assertIsInstance(self.board.get_piece(1, 0), Pawn)
+        self.assertEqual(self.board.get_piece(1, 0).get_color, "BLACK")
+        self.assertIsInstance(self.board.get_piece(6, 0), Pawn)
+        self.assertEqual(self.board.get_piece(6, 0).get_color, "WHITE")
+
+    def test_get_piece(self):
+        # Verificar que get_piece devuelve la pieza correcta
+        rook = self.board.get_piece(0, 0)
+        self.assertIsInstance(rook, Rook)
+        self.assertEqual(rook.get_color, "BLACK")
+
+        pawn = self.board.get_piece(6, 0)
+        self.assertIsInstance(pawn, Pawn)
+        self.assertEqual(pawn.get_color, "WHITE")
+
+        pawn = self.board.get_piece(1, 0)
+        self.assertIsInstance(pawn, Pawn)
+        self.assertEqual(pawn.get_color, "BLACK")
+
+    def test_set_piece(self):
+        # Verificar que se puede colocar una pieza en una posición específica
+        new_pawn = Pawn("BLACK", self.board)
+        self.board.set_piece(4, 4, new_pawn)
+        piece = self.board.get_piece(4, 4)
+        self.assertIsInstance(piece, Pawn)
+        self.assertEqual(piece.get_color, "BLACK")
+
+    def test_move(self):
+        board = Board(for_test=True)
+        rook = Rook(color='BLACK', board=board)
+        board.set_piece(0, 0, rook)
+
+        board.move(
+            from_row=0,
+            from_col=0,
+            to_row=0,
+            to_col=1,
+        )
+
+        self.assertIsInstance(
+            board.get_piece(0, 1),
+            Rook,
+        )
+        self.assertEqual(
+            str(board),
+            (
+                " ♖      \n"
+                "        \n"
+                "        \n"
+                "        \n"
+                "        \n"
+                "        \n"
+                "        \n"
+                "        \n"
+            )
+        )
+
+    def test_get_piece_out_of_range(self):
+        board = Board(for_test=True)
+
+        with self.assertRaises(OutOfBoard) as exc:
+            board.get_piece(10, 10)
+
+        self.assertEqual(
+            exc.exception.message,
+            "La posicion indicada se encuentra fuera del tablero"
+        )
+
+if __name__ == '__main__':
+    unittest.main()