+++ CHANGELOG.md
@@ -1,6 +1,133 @@
-# Changelog
+# CHANGELOG
 
-## [0.1] - 13 / 08 / 2024
+## [0.18] - 10 / 09 / 2024 - Exceptions
+
+### Added
+
+- Movimientos horizontales de la torre 
+
+
+## [0.17] - 04 / 09 / 2024 - Exceptions
+
+### Added
+
+- Excepciones básicas InvalidMove, InvalidTurn EmptyPosition
+
+## [0.16] - 03 / 09 / 2024 -  
+
+### fixed
+
+- Color de las piezas 
+
+### added
+
+- Funcion de movimiento al board
+
+## [0.15] - 02 / 09 / 2024 - Tests
+
+### Changed
+
+- Test movimientos en clase board
+- Test movimientos invalidos en board
+- Test movimiento piezas (rook) en board
+
+## [0.14] - 31 / 08 / 2024 -  Tests
+
+### Changed
+
+- Test movimientos en clase board
+
+
+## [0.13] - 30 / 08 / 2024 -  
+
+### Changed
+
+- Modificaciones movimientos invalidos y validacion de coords (chess)
+- Excepciones
+
+## [0.12] - 29 / 08 / 2024 -  
+
+### Changed
+
+- Cambios en clase piezas
+- Movimientos clase torre
+
+### Added
+
+- Funciones en clase board
+
+## [0.11] - 25 / 08 / 2024 - Test 
+
+### Added
+
+- Tests de la torre
+
+### fixed
+- Corrección de tests del tablero
+
+## [0.10] - 24 / 08 / 2024 - Board -Test
+
+### Added
+
+- Test Board movimientos
+- Posiciones en el tablero de todas las fichas
+
+## [0.9] - 23 / 08 / 2024 - Alfil
+
+### Added
+
+- Versión inicial de los movimientos del alfil
+
+## [0.8] - 22 / 08 / 2024 - Peón
+
+### added
+
+- Movimiento diagonal del peón
+
+## [0.7] - 21 / 08 / 2024 - Peón
+
+### Changed
+
+- Principales movimientos del peón 
+- Condiciones validas del movimiento
+
+## [0.6] - 20 / 08 / 2024 - Modificaciones
+
+### Added
+
+- Correcciones de variables
+- Tests del CLI
+
+## [0.5] - 19 / 08 / 2024 - Torre
+
+### Added
+
+- Primer versión del movimiento horizontal y vertical de la torre
+
+## [0.4] - 18 / 08 / 2024 - Torre
+
+### Added
+
+- Condiciones de movimiento 
+- Movimiento horizontal
+
+## [0.3] - 16 / 08 / 2024
 
 ## Agregado
+    - Mas funciones utilizando excepciones 
+
+
+## [0.2] - 15 / 08 / 2024
+
+## Agregado
+    - Funcion utilizando excepciones para validar que las coordenadas esten dentro del rango permitido
+
+
+## [0.1] - 14 / 08 / 2024
+
+## Agregado
+
+    - Creacion de clases de cada pieza
+    - Creacion del tablero 
+    - Creacion del chess y cli